+++ CHANGELOG.md
@@ -9,6 +9,7 @@
   - Fetch underlying urls
   - Clean up lock file if `bogrep` is aborted
   - Add `SelectSource` and `ReadSource` traits
+  - Add `PlistReader` and `SafariReader`
 - changed
   - Update to rust 1.75
   - Fix duplicate cache files for `bogrep fetch --urls`